Make js-hidden display !important to override.
Any usage of this class should be in a case where it is totally managed
by the javascript.  In that case, it should be as difficult as possible
to override it- if you want to display it, explicitly call the show()
method again on the element, don't try and use CSS to override.