Only modify text node if translation is different.
Previously, we would always change a text node to the result
of the translation method. This could and did result in tags
being stripped (such as <strong>), since they can show up as
text nodes, even though they have markup. By only changing
when the node text is an exact match, this issue is avoided.